fix: some error with path following with modifablecmd
not still exactly sure whyy??? but separately scheduling the trajectory command at the start of the modifiable command fixed a problem of the robot not driving in modifiablecmd. all commands are supposed to be run turn by turn sequentially, but when trajectory following command was run in parallel with the modifiablecmd (because modifiablecmd served as a wrapper for trajectory following cmd) majority the chassisspeeds calculated by the trajcmd that was sent to the drivetrain was lost. likely has to do with some weird timing issues with how commands interact. but it works now yay :D